remove bootstrap logic and pass app to every controller
This gives a much cleaner interface for the kernels. Still there is a
confusion within the getHandler method. If a class is in the controller
namespace it gets another argument (the Application).

While I don't think that it is wrong to inject the dependency injection
container to an instance it should still be in developers hand. A
controller may need other dependencies as well and the action is not
necessary in the constructor.